Thanks to karsten for sorting out my auto* issues, we can now detect where atlas is correctly.
Implemented most of the mind<->body interaction required for Character(),
and most of the operation handling for Thing(). Entities should now be
able to do stuff in the world.

Left to do, I need to work out how to implement the set operation,
and then work out a mechanism for returning multiple operation from
a function.

Al