* Cleaned up problems with mode in move operations, and dealt 	  with stopping when velocity is set to zero.
Movement now seems to work in uclient properlly. I have not yet been able
to test destination based movement in uclient, as uclient does not yet
implement right-click movement. Rounding errors in uclient also lead
to it sending move operations with very small velocities, which result
in the character walking on the spot. There are still some things to
be ironed out with face attribute.

Al