Make the poll structure static
I feel really embarrassed about this one. It should never have worked without
this made static, as this structure pointer was used as a method table by
the library, and would have been full of stack junk. Even now it is completely
unsafe if the server has two instances of the object.