Make clearTask wrapper safe
There is an assert in Character::clearTask to ensure that it is only
ever called when there is no task at shutdown. In order to ensure this is
safe in the Python API it is necessary to ensure it is never called
when there is no task, so this change adds a check to the wrapper.