Delete entity script when destroy is called
The C++ object representing this object may persist for a while as we
need it later for broadcasting ops from it as it was destroyed. Ops
in the queue from it will also hold a reference. The Python object should
be deleted right away though, as nothing else should hold a reference, and
things with weak references should find out immediately that it is gone.