Fill out class used for courses of paths and rivers
This class implements the wfmath shape concept, but does so with
a different set of template args, so can only be used by other templates
which use the concept as a partial specialization. This is not
legal in C++ 98 so will only really be possible once we have a solid
C++ 11 compiler.