Fix memory leak in AreaProperty
The property was not deleting the shape object it owned, but adding the
delete was not enough. When the property is copied, it is necessary to copy
the shape object. In order to verify, add an integration test which
simulates the normal scenario of copying properites, just like in the
StatisticsPropertyintegration test.