Alter how proximity is calculated.
We now try to see if the entity performing the task can reach the
entity which is acted upon. This is independent of the size of the
entity; the only thing that's of interest is the distance between the
edges of the two entities and the reach of a standard human.