Fix issue with world internal refno sent to client
When an Update op is generated it will contain a refno as generated by
the server.
However, when connected as a Creator client _all_ operations sent to the
entity will also be sent to the mind. The Update ops will thus collide
with the serialno/refno pairs as created by the client. To avoid this we
need to remove the refno.
Perhaps Update should use something else than refno/serialno for
matching?