Added "cyexport" tool for exporting.
This uses the same exporting code as Ember, and allows for export of
both entities, minds and types. The code in EntityExporterBase.* should
be kept synchronized with the Ember version.
This is provided as a new command, instead of added to the existing
cycmd executable. This is by design, to make it easier for users. Export
and import of entities are one of the most basic functions, especially
for any new user, and should therefore be provided as easy as
possible.