Allow instant deletion of entities through world.
If a Delete op is sent to the World with an arg with an id set the
entity with that id is instantly deleted.
This provides an override mechanism whereby an entity can be deleted
directly, by an administrator. This in contrast to the regular delete
flow where a Delete op is sent to the entity itself, and thus can be
intercepted and cancelled.