Present entity references through a map.
This changes any entity references to not present the data as a regular
string. Instead any entity reference is presented as a map, with one
entry named "$eid".
This allows clients to know what values are references to other
entities, and act upon that. This wasn't possible previously when a
simple string was used instead.
Note that those properties that implicitly refer to other entities, i.e.
"contains" and "loc" still use strings. This is because these are known
by the entity as entity referring entities.