Handle case with archetypes in the spawn.
If an archetype is specified in the spawn, but the user sends a regular
Entity definition, we need to move the attributes defined by the user to
the first entity in the archetype. This is because once the spawn has
operated on the entity, it will also be of type "archetype" and any
extra properties in the root entity will be ignored by the archetype
factory.