Check if world already is populated first.
We first check if there's any non-transient entity apart from the root
world. If so, the world is already is populated and we won't import;
unless the user supplied either the flag "--clean" which will reset the
world, or "--merge" which will try to merge the entities in the
specified file with the content of the server. Note that the "merge"
strategy in most cases will create duplicates of any existing entity.