Fix memory leak issue.
When an entity containing the respawnable property and being controlled
by an external mind is left disconnected for more than an hour the
server will enter into a state where it for each operation sent to the
entity will try to delete it. However, since it can't be deleted, and
each delete attempt actually will result in Move and Sight ops being
created, the result is a never ending build up of more and more ops,
with the end result being memory exhaustion and server crash.