Updated tests.
Started initial work on providing stubs, so that we finally can start to
reduce the amount of code needed for the tests. Some tests are using the
new stubs, but the plan is to continously move more and more code from
the tests over to the stub as they are updated.