Don't send appear ops when moving within range.
This is handled by the Move op that is broadcasted anyway. Any client
which receives a Move op from an entity that's previously not visible
knows that it's an entity that now has appeared within its sight, and
should handle it as an appear op.
Note that this breaks the protocol for older clients.