Workaround for issue with pthread.
For some reason, when using the "gold" linker, pthread isn't dynamically
linked to as it should. The result is a segfault in std::locale at
startup.
By including an object which makes a side effect free call to a pthread
function we're making sure that pthread gets dynamically linked as it
should.
This might break on systems that aren't using pthreads, but I'm
currently just focusing on getting the tests not to segfault.