Interim fix for invalid knowledge crash.
If an entity is "Unseen" it's removed from the map. This will also
delete the entity since the Python bindings don't do any refcounting.
If there's any knowledge or goal referring to it there will be a
segfault.
As an interim solution we'll therefore disable the del method for now,
until we've found a good solution for this issue.