Rewrote how plant fruiting works.
The object was to make it easier to follow, and to also remove fruit
spamming.
The attribute "fruitsMax" now denotes the maximum number of fruits in
the plant at any given time.
"fruitsChance" now denotes a percentage, i.e. [0, 100]. The closer to
100 the greater the chance of a fruit being created or dropped each
tick.
The existing fruiting entities were also altered to make them produce
less fruits, to prevent spamming.