Don't send type attributes.
This is a rather big change for the protocol, but by now there shouldn't
be any clients left that can't handle this.
Basically, since the client already has the type and knows the default
properties from that, there's no need to resend them for each entity.
This should decrease the amount of data needed to be sent from the
server.

Note that there might be some lingering issues with the AI code; that
has to be investigated further.