(WIP) Rewrite filter tests using TestBase class
This is a fairly complex change with a simple goal.
Instead of holding all test cases in the main function,
we reorganize it into a proper test suit.
The tests are groupped up into appropriate test methods.
This should improve the clarity of test cases and ease
addition of new test cases.