Fix overlooked bugs in existing providers tests
These bugs were overlooked because even though the existing
ASSERT macros in the test suite seem to be clearer and
more informative, they do not report a failure of the test
suite even if a given test fails. They only make a not in the log.
For now, we will use built-in assert to make sure that
failing tests are caught.