Provide immorality feature through scripts.
This approach will be taken for most properties.
The main problem with having logic in properties is that the order in
which the code is executed is random. This makes it too hard to reason
about the system. Instead we prefer the deterministic nature of having
scripts in a specific order.