Introduce the "properties" attribute for types.
When defining types you define the values that are to be applied to the
type in the "attributes" element. These can be a combination of
"default", "append", "prepend" and "subtract".
On the server, the "attributes" will be calculated for each type, and
the results put in the "properties" element.
This is the element that the clients look at. For regular clients only
the "properties" element will be sent; only admin clients will be
allowed to see the underlying "attributes" element.