Introduce non-recursive "contains".
When checking if something is in inventory we don't want to use the
"contains_recursive" function, instead we just want to check in the
first level. That's why the "contains" function now is added.
As a consequence the keyword for checking if something is included in a
list has now been changed from "contains" to "includes".
I.e. "[1,2,3] includes 1"