Added Container domain.
A container domain is meant to be used for containers, i.e. entities
that contains other entities and that cannot be inspected unless the
inspecting entity is within reachable distance.

It's not totally clear how we should handle appear/disappear on the
client side yet. We want a system where it's clear to clients that they
need to get closer in order to see the contents.