Change protocol so location data is sent as standard Set op.
Previously we only allowed updates to location data (pos, velocity etc.) to be sent as
a Sight of a Move op. However, there were no real reason for this. Now we sent the changes
as a standard Set op.
The client still needs to send a Move op in order to move though.