Added concept of "modifers" and "ModifyProperty".
This is a way in which a child entity can affect its parent entity. A typical modification would
be to have a piece of armor affect the wearer's defense values.
Optionally Entity Filters can be applied as constraints. In the armor case it would mean that the
modification only would be active if the armor was worn.