Added "close_container" op for closing containers.
This is made into a core C++ feature, mainly because we don't want gameplay where
the user is unable to close containers they have opened.
Preferably it should have been in the rules part, but I consider it too
important to not leave opened containers without ability to close them.