Refactor how spawns works.
We need to remove all code from C++ which handles spawning new entities into the world
and put it in Python scripts on the entities.
When an entity contains the "__spawn" property it's registered as a possible "spawn point".
Clients are allowed to send Create op tot he spawn points, which in turn rely on there being
Python scripts which intercepts these and validate them for creation of new characters.