Simplify water body movement handling.
Since we don't expect water bodies to move much in game we optimize for this
by not having extra lookups in place. Instead we just brute force go through
all entities when a water body move (which we expect them pretty much never to do).

Also removed the "floats" handling. If an entity should be placed on top of a
water body it needs to be planted on it. There's work remaining to make sure that
free moving entities that naturally float can do so.