Send a more correct version of the entity if in "archive" mode.
If the client want to store the entity so it later on can be restored it's easier
if we don't send any ephemeral props, and we send the base value for any modified prop.
We could achieve the same result by looking at the "_modifiers" prop, but this is just
simpler.